backend-ml/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.3.2
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
